requirements.txt
anthropic>=0.40.0
openai>=1.30.0
yfinance>=0.2.40
requests>=2.31.0
feedgen>=1.0.0
pydub>=0.25.1
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.130.0
beautifulsoup4>=4.12.0
